Fix master_failure_mode=error_on_write
The read-only error was not sent if the query that was attempted was a
read and no valid server was found. Moved the routing failure handling to
its own method to clarify the code in route_single_stmt and made the
create_readonly_error function a part of send_readonly_error.

This commit fixes the rwsplit_readonly test.